+++ src/gufo/thor/services/mongo.py
@@ -10,7 +10,11 @@ Attributes:
     mongo: mongo service singleton.
 """
 
+# NOC modules
+from typing import List, Optional
+
 # Gufo Thor modules
+from ..config import Config, ServiceConfig
 from .base import BaseService, ComposeDependsCondition
 
 
@@ -31,6 +35,15 @@ class MongoService(BaseService):
     compose_volumes = ["mongo_data:/data/db", "backup:/data/backup"]
     compose_volumes_config = {"mongo_data": {}}
     service_discovery = {"mongo": 27017}
+
+    def get_compose_ports(
+        self: "MongoService", config: Config, svc: Optional[ServiceConfig]
+    ) -> Optional[List[str]]:
+        """Expose port."""
+        r = super().get_compose_ports(config, svc) or []
+        if config.expose.mongo:
+            r.append(config.expose.mongo.docker_compose_port(27017))
+        return r if r else None
 
 
 mongo = MongoService()